pyproject.toml
[tool.poetry]
name = "gallery"
version = "0.1.0"
description = ""
authors = ["Khasanjon Mamadaliyev <khasanjon.eng@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.2"
python-dotenv = "^1.0.0"
drf-yasg = "^1.21.6"
djangorestframework = "^3.14.0"
moviepy = "^1.0.3"
djangorestframework-simplejwt = "^5.2.2"
psycopg2-binary = "^2.9.6"
django-redis = "^5.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
